refactor: added more fns to utils
