+++ test/suites/zombie_tanssi_relay_unneeded_para/test_zombie_tanssi_relay_unneeded_para.ts
@@ -3,7 +3,15 @@ import { MIN_GAS_PRICE, customWeb3Request, generateKeyringPair } from "@moonwall
 import { type ApiPromise, Keyring } from "@polkadot/api";
 import type { Signer } from "ethers";
 import fs from "node:fs/promises";
-import { createTransfer, getHeaderFromRelay, signAndSendAndInclude, waitSessions, waitUntilEthTxIncluded } from "utils";
+import {
+    checkLogsNotExist,
+    createTransfer,
+    getHeaderFromRelay,
+    getTmpZombiePath,
+    signAndSendAndInclude,
+    waitSessions,
+    waitUntilEthTxIncluded,
+} from "utils";
 
 describeSuite({
     id: "ZOMBIETANSSIR01",
@@ -286,32 +294,3 @@ describeSuite({
         });
     },
 });
-
-// Read log file path and check that none of the specified logs are found.
-// Only supports single-line logs.
-async function checkLogsNotExist(logFilePath: string, logs: string[]): Promise<void> {
-    const fileContent = await fs.readFile(logFilePath, "utf8");
-    const lines = fileContent.split("\n");
-
-    for (let i = 0; i < lines.length; i++) {
-        for (const log of logs) {
-            if (lines[i].includes(log)) {
-                // In case any log is found, show some context around the found log
-                const contextSize = 3;
-                const contextStart = Math.max(0, i - contextSize);
-                const contextEnd = Math.min(lines.length - 1, i + contextSize);
-                const contextLines = lines.slice(contextStart, contextEnd + 1);
-                const contextStr = contextLines.join("\n");
-
-                expect.fail(
-                    `Log entry '${log}' was found in the log file.\nContext around the found log:\n${contextStr}`
-                );
-            }
-        }
-    }
-}
-
-/// Returns the /tmp/zombie-52234... path
-function getTmpZombiePath() {
-    return process.env.MOON_ZOMBIE_DIR;
-}
