+++ test/utils/typescript.ts
@@ -1,0 +1,3 @@
+export async function sleep(ms: number) {
+    await new Promise((resolve) => setTimeout(resolve, ms));
+}
